Isolate `pr view` command
