+++ command/pr.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -15,6 +14,7 @@ import (
 	"github.com/cli/cli/context"
 	"github.com/cli/cli/git"
 	"github.com/cli/cli/internal/ghrepo"
+	"github.com/cli/cli/pkg/cmd/pr/shared"
 	"github.com/cli/cli/pkg/cmdutil"
 	"github.com/cli/cli/pkg/prompt"
 	"github.com/cli/cli/pkg/text"
@@ -45,9 +45,6 @@ func init() {
 	prListCmd.Flags().StringP("base", "B", "", "Filter by base branch")
 	prListCmd.Flags().StringSliceP("label", "l", nil, "Filter by labels")
 	prListCmd.Flags().StringP("assignee", "a", "", "Filter by assignee")
-
-	prCmd.AddCommand(prViewCmd)
-	prViewCmd.Flags().BoolP("web", "w", false, "Open a pull request in the browser")
 }
 
 var prCmd = &cobra.Command{
@@ -83,17 +80,6 @@ var prStatusCmd = &cobra.Command{
 	Short: "Show status of relevant pull requests",
 	Args:  cmdutil.NoArgsQuoteReminder,
 	RunE:  prStatus,
-}
-var prViewCmd = &cobra.Command{
-	Use:   "view [<number> | <url> | <branch>]",
-	Short: "View a pull request",
-	Long: `Display the title, body, and other information about a pull request.
-
-Without an argument, the pull request that belongs to the current branch
-is displayed.
-
-With '--web', open the pull request in a web browser instead.`,
-	RunE: prView,
 }
 var prCloseCmd = &cobra.Command{
 	Use:   "close {<number> | <url> | <branch>}",
@@ -303,7 +289,7 @@ func prList(cmd *cobra.Command, args []string) error {
 		if table.IsTTY() {
 			prNum = "#" + prNum
 		}
-		table.AddField(prNum, nil, colorFuncForPR(pr))
+		table.AddField(prNum, nil, shared.ColorFuncForPR(pr))
 		table.AddField(replaceExcessiveWhitespace(pr.Title), nil, nil)
 		table.AddField(pr.HeadLabel(), nil, utils.Cyan)
 		if !table.IsTTY() {
@@ -317,69 +303,6 @@ func prList(cmd *cobra.Command, args []string) error {
 	}
 
 	return nil
-}
-
-func prStateTitleWithColor(pr api.PullRequest) string {
-	prStateColorFunc := colorFuncForPR(pr)
-	if pr.State == "OPEN" && pr.IsDraft {
-		return prStateColorFunc(strings.Title(strings.ToLower("Draft")))
-	}
-	return prStateColorFunc(strings.Title(strings.ToLower(pr.State)))
-}
-
-func colorFuncForPR(pr api.PullRequest) func(string) string {
-	if pr.State == "OPEN" && pr.IsDraft {
-		return utils.Gray
-	}
-	return colorFuncForState(pr.State)
-}
-
-// colorFuncForState returns a color function for a PR/Issue state
-func colorFuncForState(state string) func(string) string {
-	switch state {
-	case "OPEN":
-		return utils.Green
-	case "CLOSED":
-		return utils.Red
-	case "MERGED":
-		return utils.Magenta
-	default:
-		return nil
-	}
-}
-
-func prView(cmd *cobra.Command, args []string) error {
-	ctx := contextForCommand(cmd)
-
-	apiClient, err := apiClientForContext(ctx)
-	if err != nil {
-		return err
-	}
-
-	web, err := cmd.Flags().GetBool("web")
-	if err != nil {
-		return err
-	}
-
-	pr, _, err := prFromArgs(ctx, apiClient, cmd, args)
-	if err != nil {
-		return err
-	}
-	openURL := pr.URL
-
-	if web {
-		if connectedToTerminal(cmd) {
-			fmt.Fprintf(cmd.ErrOrStderr(), "Opening %s in your browser.\n", openURL)
-		}
-		return utils.OpenInBrowser(openURL)
-	}
-
-	if connectedToTerminal(cmd) {
-		out := colorableOut(cmd)
-		return printHumanPrPreview(out, pr)
-	}
-
-	return printRawPrPreview(cmd.OutOrStdout(), pr)
 }
 
 func prClose(cmd *cobra.Command, args []string) error {
@@ -646,78 +569,6 @@ func prStateWithDraft(pr *api.PullRequest) string {
 	return pr.State
 }
 
-func printRawPrPreview(out io.Writer, pr *api.PullRequest) error {
-	reviewers := prReviewerList(*pr)
-	assignees := prAssigneeList(*pr)
-	labels := prLabelList(*pr)
-	projects := prProjectList(*pr)
-
-	fmt.Fprintf(out, "title:\t%s\n", pr.Title)
-	fmt.Fprintf(out, "state:\t%s\n", prStateWithDraft(pr))
-	fmt.Fprintf(out, "author:\t%s\n", pr.Author.Login)
-	fmt.Fprintf(out, "labels:\t%s\n", labels)
-	fmt.Fprintf(out, "assignees:\t%s\n", assignees)
-	fmt.Fprintf(out, "reviewers:\t%s\n", reviewers)
-	fmt.Fprintf(out, "projects:\t%s\n", projects)
-	fmt.Fprintf(out, "milestone:\t%s\n", pr.Milestone.Title)
-
-	fmt.Fprintln(out, "--")
-	fmt.Fprintln(out, pr.Body)
-
-	return nil
-}
-
-func printHumanPrPreview(out io.Writer, pr *api.PullRequest) error {
-	// Header (Title and State)
-	fmt.Fprintln(out, utils.Bold(pr.Title))
-	fmt.Fprintf(out, "%s", prStateTitleWithColor(*pr))
-	fmt.Fprintln(out, utils.Gray(fmt.Sprintf(
-		" • %s wants to merge %s into %s from %s",
-		pr.Author.Login,
-		utils.Pluralize(pr.Commits.TotalCount, "commit"),
-		pr.BaseRefName,
-		pr.HeadRefName,
-	)))
-	fmt.Fprintln(out)
-
-	// Metadata
-	if reviewers := prReviewerList(*pr); reviewers != "" {
-		fmt.Fprint(out, utils.Bold("Reviewers: "))
-		fmt.Fprintln(out, reviewers)
-	}
-	if assignees := prAssigneeList(*pr); assignees != "" {
-		fmt.Fprint(out, utils.Bold("Assignees: "))
-		fmt.Fprintln(out, assignees)
-	}
-	if labels := prLabelList(*pr); labels != "" {
-		fmt.Fprint(out, utils.Bold("Labels: "))
-		fmt.Fprintln(out, labels)
-	}
-	if projects := prProjectList(*pr); projects != "" {
-		fmt.Fprint(out, utils.Bold("Projects: "))
-		fmt.Fprintln(out, projects)
-	}
-	if pr.Milestone.Title != "" {
-		fmt.Fprint(out, utils.Bold("Milestone: "))
-		fmt.Fprintln(out, pr.Milestone.Title)
-	}
-
-	// Body
-	if pr.Body != "" {
-		fmt.Fprintln(out)
-		md, err := utils.RenderMarkdown(pr.Body)
-		if err != nil {
-			return err
-		}
-		fmt.Fprintln(out, md)
-	}
-	fmt.Fprintln(out)
-
-	// Footer
-	fmt.Fprintf(out, utils.Gray("View this pull request on GitHub: %s\n"), pr.URL)
-	return nil
-}
-
 func prReady(cmd *cobra.Command, args []string) error {
 	ctx := contextForCommand(cmd)
 	apiClient, err := apiClientForContext(ctx)
@@ -746,182 +597,6 @@ func prReady(cmd *cobra.Command, args []string) error {
 	fmt.Fprintf(colorableErr(cmd), "%s Pull request #%d is marked as \"ready for review\"\n", utils.Green("✔"), pr.Number)
 
 	return nil
-}
-
-// Ref. https://developer.github.com/v4/enum/pullrequestreviewstate/
-const (
-	requestedReviewState        = "REQUESTED" // This is our own state for review request
-	approvedReviewState         = "APPROVED"
-	changesRequestedReviewState = "CHANGES_REQUESTED"
-	commentedReviewState        = "COMMENTED"
-	dismissedReviewState        = "DISMISSED"
-	pendingReviewState          = "PENDING"
-)
-
-type reviewerState struct {
-	Name  string
-	State string
-}
-
-// colorFuncForReviewerState returns a color function for a reviewer state
-func colorFuncForReviewerState(state string) func(string) string {
-	switch state {
-	case requestedReviewState:
-		return utils.Yellow
-	case approvedReviewState:
-		return utils.Green
-	case changesRequestedReviewState:
-		return utils.Red
-	case commentedReviewState:
-		return func(str string) string { return str } // Do nothing
-	default:
-		return nil
-	}
-}
-
-// formattedReviewerState formats a reviewerState with state color
-func formattedReviewerState(reviewer *reviewerState) string {
-	state := reviewer.State
-	if state == dismissedReviewState {
-		// Show "DISMISSED" review as "COMMENTED", since "dimissed" only makes
-		// sense when displayed in an events timeline but not in the final tally.
-		state = commentedReviewState
-	}
-	stateColorFunc := colorFuncForReviewerState(state)
-	return fmt.Sprintf("%s (%s)", reviewer.Name, stateColorFunc(strings.ReplaceAll(strings.Title(strings.ToLower(state)), "_", " ")))
-}
-
-// prReviewerList generates a reviewer list with their last state
-func prReviewerList(pr api.PullRequest) string {
-	reviewerStates := parseReviewers(pr)
-	reviewers := make([]string, 0, len(reviewerStates))
-
-	sortReviewerStates(reviewerStates)
-
-	for _, reviewer := range reviewerStates {
-		reviewers = append(reviewers, formattedReviewerState(reviewer))
-	}
-
-	reviewerList := strings.Join(reviewers, ", ")
-
-	return reviewerList
-}
-
-// Ref. https://developer.github.com/v4/union/requestedreviewer/
-const teamTypeName = "Team"
-
-const ghostName = "ghost"
-
-// parseReviewers parses given Reviews and ReviewRequests
-func parseReviewers(pr api.PullRequest) []*reviewerState {
-	reviewerStates := make(map[string]*reviewerState)
-
-	for _, review := range pr.Reviews.Nodes {
-		if review.Author.Login != pr.Author.Login {
-			name := review.Author.Login
-			if name == "" {
-				name = ghostName
-			}
-			reviewerStates[name] = &reviewerState{
-				Name:  name,
-				State: review.State,
-			}
-		}
-	}
-
-	// Overwrite reviewer's state if a review request for the same reviewer exists.
-	for _, reviewRequest := range pr.ReviewRequests.Nodes {
-		name := reviewRequest.RequestedReviewer.Login
-		if reviewRequest.RequestedReviewer.TypeName == teamTypeName {
-			name = reviewRequest.RequestedReviewer.Name
-		}
-		reviewerStates[name] = &reviewerState{
-			Name:  name,
-			State: requestedReviewState,
-		}
-	}
-
-	// Convert map to slice for ease of sort
-	result := make([]*reviewerState, 0, len(reviewerStates))
-	for _, reviewer := range reviewerStates {
-		if reviewer.State == pendingReviewState {
-			continue
-		}
-		result = append(result, reviewer)
-	}
-
-	return result
-}
-
-// sortReviewerStates puts completed reviews before review requests and sorts names alphabetically
-func sortReviewerStates(reviewerStates []*reviewerState) {
-	sort.Slice(reviewerStates, func(i, j int) bool {
-		if reviewerStates[i].State == requestedReviewState &&
-			reviewerStates[j].State != requestedReviewState {
-			return false
-		}
-		if reviewerStates[j].State == requestedReviewState &&
-			reviewerStates[i].State != requestedReviewState {
-			return true
-		}
-
-		return reviewerStates[i].Name < reviewerStates[j].Name
-	})
-}
-
-func prAssigneeList(pr api.PullRequest) string {
-	if len(pr.Assignees.Nodes) == 0 {
-		return ""
-	}
-
-	AssigneeNames := make([]string, 0, len(pr.Assignees.Nodes))
-	for _, assignee := range pr.Assignees.Nodes {
-		AssigneeNames = append(AssigneeNames, assignee.Login)
-	}
-
-	list := strings.Join(AssigneeNames, ", ")
-	if pr.Assignees.TotalCount > len(pr.Assignees.Nodes) {
-		list += ", …"
-	}
-	return list
-}
-
-func prLabelList(pr api.PullRequest) string {
-	if len(pr.Labels.Nodes) == 0 {
-		return ""
-	}
-
-	labelNames := make([]string, 0, len(pr.Labels.Nodes))
-	for _, label := range pr.Labels.Nodes {
-		labelNames = append(labelNames, label.Name)
-	}
-
-	list := strings.Join(labelNames, ", ")
-	if pr.Labels.TotalCount > len(pr.Labels.Nodes) {
-		list += ", …"
-	}
-	return list
-}
-
-func prProjectList(pr api.PullRequest) string {
-	if len(pr.ProjectCards.Nodes) == 0 {
-		return ""
-	}
-
-	projectNames := make([]string, 0, len(pr.ProjectCards.Nodes))
-	for _, project := range pr.ProjectCards.Nodes {
-		colName := project.Column.Name
-		if colName == "" {
-			colName = "Awaiting triage"
-		}
-		projectNames = append(projectNames, fmt.Sprintf("%s (%s)", project.Project.Name, colName))
-	}
-
-	list := strings.Join(projectNames, ", ")
-	if pr.ProjectCards.TotalCount > len(pr.ProjectCards.Nodes) {
-		list += ", …"
-	}
-	return list
 }
 
 func prSelectorForCurrentBranch(ctx context.Context, baseRepo ghrepo.Interface) (prNumber int, prHeadRef string, err error) {
@@ -1019,7 +694,7 @@ func printPrs(w io.Writer, totalCount int, prs ...api.PullRequest) {
 				fmt.Fprint(w, utils.Green("✓ Approved"))
 			}
 		} else {
-			fmt.Fprintf(w, " - %s", prStateTitleWithColor(pr))
+			fmt.Fprintf(w, " - %s", shared.StateTitleWithColor(pr))
 		}
 
 		fmt.Fprint(w, "\n")
