+++ command/pr_test.go
@@ -8,12 +8,10 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/cli/cli/api"
 	"github.com/cli/cli/internal/run"
 	"github.com/cli/cli/pkg/httpmock"
 	"github.com/cli/cli/pkg/prompt"
 	"github.com/cli/cli/test"
-	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -484,592 +482,11 @@ func TestPRList_web(t *testing.T) {
 	eq(t, url, expectedURL)
 }
 
-func TestPRView_Preview_nontty(t *testing.T) {
-	defer stubTerminal(false)()
-	tests := map[string]struct {
-		ownerRepo       string
-		args            string
-		fixture         string
-		expectedOutputs []string
-	}{
-		"Open PR without metadata": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreview.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are from a fork\n`,
-				`state:\tOPEN\n`,
-				`author:\tnobody\n`,
-				`labels:\t\n`,
-				`assignees:\t\n`,
-				`reviewers:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-				`blueberries taste good`,
-			},
-		},
-		"Open PR with metadata by number": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewWithMetadataByNumber.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are from a fork\n`,
-				`reviewers:\t2 \(Approved\), 3 \(Commented\), 1 \(Requested\)\n`,
-				`assignees:\tmarseilles, monaco\n`,
-				`labels:\tone, two, three, four, five\n`,
-				`projects:\tProject 1 \(column A\), Project 2 \(column B\), Project 3 \(column C\), Project 4 \(Awaiting triage\)\n`,
-				`milestone:\tuluru\n`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Open PR with reviewers by number": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewWithReviewersByNumber.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are from a fork\n`,
-				`state:\tOPEN\n`,
-				`author:\tnobody\n`,
-				`labels:\t\n`,
-				`assignees:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-				`reviewers:\tDEF \(Commented\), def \(Changes requested\), ghost \(Approved\), hubot \(Commented\), xyz \(Approved\), 123 \(Requested\), Team 1 \(Requested\), abc \(Requested\)\n`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Open PR with metadata by branch": {
-			ownerRepo: "master",
-			args:      "pr view blueberries",
-			fixture:   "../test/fixtures/prViewPreviewWithMetadataByBranch.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are a good fruit`,
-				`state:\tOPEN`,
-				`author:\tnobody`,
-				`assignees:\tmarseilles, monaco\n`,
-				`labels:\tone, two, three, four, five\n`,
-				`projects:\tProject 1 \(column A\), Project 2 \(column B\), Project 3 \(column C\)\n`,
-				`milestone:\tuluru\n`,
-				`blueberries taste good`,
-			},
-		},
-		"Open PR for the current branch": {
-			ownerRepo: "blueberries",
-			args:      "pr view",
-			fixture:   "../test/fixtures/prView.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are a good fruit`,
-				`state:\tOPEN`,
-				`author:\tnobody`,
-				`assignees:\t\n`,
-				`labels:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Open PR wth empty body for the current branch": {
-			ownerRepo: "blueberries",
-			args:      "pr view",
-			fixture:   "../test/fixtures/prView_EmptyBody.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are a good fruit`,
-				`state:\tOPEN`,
-				`author:\tnobody`,
-				`assignees:\t\n`,
-				`labels:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-			},
-		},
-		"Closed PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewClosedState.json",
-			expectedOutputs: []string{
-				`state:\tCLOSED\n`,
-				`author:\tnobody\n`,
-				`labels:\t\n`,
-				`assignees:\t\n`,
-				`reviewers:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Merged PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewMergedState.json",
-			expectedOutputs: []string{
-				`state:\tMERGED\n`,
-				`author:\tnobody\n`,
-				`labels:\t\n`,
-				`assignees:\t\n`,
-				`reviewers:\t\n`,
-				`projects:\t\n`,
-				`milestone:\t\n`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Draft PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewDraftState.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are from a fork\n`,
-				`state:\tDRAFT\n`,
-				`author:\tnobody\n`,
-				`labels:`,
-				`assignees:`,
-				`projects:`,
-				`milestone:`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-		"Draft PR by branch": {
-			ownerRepo: "master",
-			args:      "pr view blueberries",
-			fixture:   "../test/fixtures/prViewPreviewDraftStatebyBranch.json",
-			expectedOutputs: []string{
-				`title:\tBlueberries are a good fruit\n`,
-				`state:\tDRAFT\n`,
-				`author:\tnobody\n`,
-				`labels:`,
-				`assignees:`,
-				`projects:`,
-				`milestone:`,
-				`\*\*blueberries taste good\*\*`,
-			},
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			initBlankContext("", "OWNER/REPO", tc.ownerRepo)
-			http := initFakeHTTP()
-			http.StubRepoResponse("OWNER", "REPO")
-			http.Register(httpmock.GraphQL(`query PullRequest(ByNumber|ForBranch)\b`), httpmock.FileResponse(tc.fixture))
-
-			output, err := RunCommand(tc.args)
-			if err != nil {
-				t.Errorf("error running command `%v`: %v", tc.args, err)
-			}
-
-			eq(t, output.Stderr(), "")
-
-			test.ExpectLines(t, output.String(), tc.expectedOutputs...)
-		})
-	}
-}
-
-func TestPRView_Preview(t *testing.T) {
-	defer stubTerminal(true)()
-	tests := map[string]struct {
-		ownerRepo       string
-		args            string
-		fixture         string
-		expectedOutputs []string
-	}{
-		"Open PR without metadata": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreview.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Open.*nobody wants to merge 12 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12`,
-			},
-		},
-		"Open PR with metadata by number": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewWithMetadataByNumber.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Open.*nobody wants to merge 12 commits into master from blueberries`,
-				`Reviewers:.*2 \(.*Approved.*\), 3 \(Commented\), 1 \(.*Requested.*\)\n`,
-				`Assignees:.*marseilles, monaco\n`,
-				`Labels:.*one, two, three, four, five\n`,
-				`Projects:.*Project 1 \(column A\), Project 2 \(column B\), Project 3 \(column C\), Project 4 \(Awaiting triage\)\n`,
-				`Milestone:.*uluru\n`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12\n`,
-			},
-		},
-		"Open PR with reviewers by number": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewWithReviewersByNumber.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Reviewers:.*DEF \(.*Commented.*\), def \(.*Changes requested.*\), ghost \(.*Approved.*\), hubot \(Commented\), xyz \(.*Approved.*\), 123 \(.*Requested.*\), Team 1 \(.*Requested.*\), abc \(.*Requested.*\)\n`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12\n`,
-			},
-		},
-		"Open PR with metadata by branch": {
-			ownerRepo: "master",
-			args:      "pr view blueberries",
-			fixture:   "../test/fixtures/prViewPreviewWithMetadataByBranch.json",
-			expectedOutputs: []string{
-				`Blueberries are a good fruit`,
-				`Open.*nobody wants to merge 8 commits into master from blueberries`,
-				`Assignees:.*marseilles, monaco\n`,
-				`Labels:.*one, two, three, four, five\n`,
-				`Projects:.*Project 1 \(column A\), Project 2 \(column B\), Project 3 \(column C\)\n`,
-				`Milestone:.*uluru\n`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/10\n`,
-			},
-		},
-		"Open PR for the current branch": {
-			ownerRepo: "blueberries",
-			args:      "pr view",
-			fixture:   "../test/fixtures/prView.json",
-			expectedOutputs: []string{
-				`Blueberries are a good fruit`,
-				`Open.*nobody wants to merge 8 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/10`,
-			},
-		},
-		"Open PR wth empty body for the current branch": {
-			ownerRepo: "blueberries",
-			args:      "pr view",
-			fixture:   "../test/fixtures/prView_EmptyBody.json",
-			expectedOutputs: []string{
-				`Blueberries are a good fruit`,
-				`Open.*nobody wants to merge 8 commits into master from blueberries`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/10`,
-			},
-		},
-		"Closed PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewClosedState.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Closed.*nobody wants to merge 12 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12`,
-			},
-		},
-		"Merged PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewMergedState.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Merged.*nobody wants to merge 12 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12`,
-			},
-		},
-		"Draft PR": {
-			ownerRepo: "master",
-			args:      "pr view 12",
-			fixture:   "../test/fixtures/prViewPreviewDraftState.json",
-			expectedOutputs: []string{
-				`Blueberries are from a fork`,
-				`Draft.*nobody wants to merge 12 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/12`,
-			},
-		},
-		"Draft PR by branch": {
-			ownerRepo: "master",
-			args:      "pr view blueberries",
-			fixture:   "../test/fixtures/prViewPreviewDraftStatebyBranch.json",
-			expectedOutputs: []string{
-				`Blueberries are a good fruit`,
-				`Draft.*nobody wants to merge 8 commits into master from blueberries`,
-				`blueberries taste good`,
-				`View this pull request on GitHub: https://github.com/OWNER/REPO/pull/10`,
-			},
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			initBlankContext("", "OWNER/REPO", tc.ownerRepo)
-			http := initFakeHTTP()
-			http.StubRepoResponse("OWNER", "REPO")
-			http.Register(httpmock.GraphQL(`query PullRequest(ByNumber|ForBranch)\b`), httpmock.FileResponse(tc.fixture))
-
-			output, err := RunCommand(tc.args)
-			if err != nil {
-				t.Errorf("error running command `%v`: %v", tc.args, err)
-			}
-
-			eq(t, output.Stderr(), "")
-
-			test.ExpectLines(t, output.String(), tc.expectedOutputs...)
-		})
-	}
-}
-
-func TestPRView_web_currentBranch(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestForBranch\b`), httpmock.FileResponse("../test/fixtures/prView.json"))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		switch strings.Join(cmd.Args, " ") {
-		case `git config --get-regexp ^branch\.blueberries\.(remote|merge)$`:
-			return &test.OutputStub{}
-		default:
-			seenCmd = cmd
-			return &test.OutputStub{}
-		}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-	eq(t, output.Stderr(), "Opening https://github.com/OWNER/REPO/pull/10 in your browser.\n")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	if url != "https://github.com/OWNER/REPO/pull/10" {
-		t.Errorf("got: %q", url)
-	}
-}
-
-func TestPRView_web_noResultsForBranch(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "blueberries")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-	http.Register(httpmock.GraphQL(`query PullRequestForBranch\b`), httpmock.FileResponse("../test/fixtures/prView_NoActiveBranch.json"))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		switch strings.Join(cmd.Args, " ") {
-		case `git config --get-regexp ^branch\.blueberries\.(remote|merge)$`:
-			return &test.OutputStub{}
-		default:
-			seenCmd = cmd
-			return &test.OutputStub{}
-		}
-	})
-	defer restoreCmd()
-
-	_, err := RunCommand("pr view -w")
-	if err == nil || err.Error() != `no open pull requests found for branch "blueberries"` {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	if seenCmd != nil {
-		t.Fatalf("unexpected command: %v", seenCmd.Args)
-	}
-}
-
-func TestPRView_web_numberArg(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "master")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-
-	http.StubResponse(200, bytes.NewBufferString(`
-	{ "data": { "repository": { "pullRequest": {
-		"url": "https://github.com/OWNER/REPO/pull/23"
-	} } } }
-	`))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		seenCmd = cmd
-		return &test.OutputStub{}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w 23")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	eq(t, url, "https://github.com/OWNER/REPO/pull/23")
-}
-
-func TestPRView_web_numberArgWithHash(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "master")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-
-	http.StubResponse(200, bytes.NewBufferString(`
-	{ "data": { "repository": { "pullRequest": {
-		"url": "https://github.com/OWNER/REPO/pull/23"
-	} } } }
-	`))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		seenCmd = cmd
-		return &test.OutputStub{}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w \"#23\"")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	eq(t, url, "https://github.com/OWNER/REPO/pull/23")
-}
-
-func TestPRView_web_urlArg(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "master")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubResponse(200, bytes.NewBufferString(`
-		{ "data": { "repository": { "pullRequest": {
-			"url": "https://github.com/OWNER/REPO/pull/23"
-		} } } }
-	`))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		seenCmd = cmd
-		return &test.OutputStub{}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w https://github.com/OWNER/REPO/pull/23/files")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	eq(t, url, "https://github.com/OWNER/REPO/pull/23")
-}
-
-func TestPRView_web_branchArg(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "master")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-
-	http.StubResponse(200, bytes.NewBufferString(`
-	{ "data": { "repository": { "pullRequests": { "nodes": [
-		{ "headRefName": "blueberries",
-		  "isCrossRepository": false,
-		  "url": "https://github.com/OWNER/REPO/pull/23" }
-	] } } } }
-	`))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		seenCmd = cmd
-		return &test.OutputStub{}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w blueberries")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	eq(t, url, "https://github.com/OWNER/REPO/pull/23")
-}
-
-func TestPRView_web_branchWithOwnerArg(t *testing.T) {
-	initBlankContext("", "OWNER/REPO", "master")
-	defer stubTerminal(true)()
-	http := initFakeHTTP()
-	http.StubRepoResponse("OWNER", "REPO")
-
-	http.StubResponse(200, bytes.NewBufferString(`
-	{ "data": { "repository": { "pullRequests": { "nodes": [
-		{ "headRefName": "blueberries",
-		  "isCrossRepository": true,
-		  "headRepositoryOwner": { "login": "hubot" },
-		  "url": "https://github.com/hubot/REPO/pull/23" }
-	] } } } }
-	`))
-
-	var seenCmd *exec.Cmd
-	restoreCmd := run.SetPrepareCmd(func(cmd *exec.Cmd) run.Runnable {
-		seenCmd = cmd
-		return &test.OutputStub{}
-	})
-	defer restoreCmd()
-
-	output, err := RunCommand("pr view -w hubot:blueberries")
-	if err != nil {
-		t.Errorf("error running command `pr view`: %v", err)
-	}
-
-	eq(t, output.String(), "")
-
-	if seenCmd == nil {
-		t.Fatal("expected a command to run")
-	}
-	url := seenCmd.Args[len(seenCmd.Args)-1]
-	eq(t, url, "https://github.com/hubot/REPO/pull/23")
-}
-
 func TestReplaceExcessiveWhitespace(t *testing.T) {
 	eq(t, replaceExcessiveWhitespace("hello\ngoodbye"), "hello goodbye")
 	eq(t, replaceExcessiveWhitespace("  hello goodbye  "), "hello goodbye")
 	eq(t, replaceExcessiveWhitespace("hello   goodbye"), "hello goodbye")
 	eq(t, replaceExcessiveWhitespace("   hello \n goodbye   "), "hello goodbye")
-}
-
-func TestPrStateTitleWithColor(t *testing.T) {
-	tests := map[string]struct {
-		pr   api.PullRequest
-		want string
-	}{
-		"Format OPEN state":              {pr: api.PullRequest{State: "OPEN", IsDraft: false}, want: "Open"},
-		"Format OPEN state for Draft PR": {pr: api.PullRequest{State: "OPEN", IsDraft: true}, want: "Draft"},
-		"Format CLOSED state":            {pr: api.PullRequest{State: "CLOSED", IsDraft: false}, want: "Closed"},
-		"Format MERGED state":            {pr: api.PullRequest{State: "MERGED", IsDraft: false}, want: "Merged"},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			got := prStateTitleWithColor(tc.pr)
-			diff := cmp.Diff(tc.want, got)
-			if diff != "" {
-				t.Fatalf(diff)
-			}
-		})
-	}
 }
 
 func TestPrClose(t *testing.T) {
