+++ pkg/cmd/pr/view/fixtures/prViewPreviewClosedState.json
