+++ pkg/cmd/pr/view/fixtures/prViewPreviewDraftState.json
