+++ pkg/cmd/pr/view/fixtures/prViewPreviewDraftStatebyBranch.json
