+++ pkg/cmd/pr/view/fixtures/prViewPreviewMergedState.json
