+++ pkg/cmd/pr/view/fixtures/prViewPreviewWithMetadataByBranch.json
