+++ pkg/cmd/pr/view/fixtures/prViewPreviewWithMetadataByNumber.json
