+++ pkg/cmd/pr/view/fixtures/prView_EmptyBody.json
