+++ pkg/cmd/pr/view/fixtures/prView_NoActiveBranch.json
